requirements.txt
fastapi
kbinxml>=2.0
pycryptodomex
python-multipart
tinydb
uvicorn[standard]
